+++ first-world/Logic/IAgeable.cs
@@ -2,7 +2,7 @@ namespace FirstWorld.Logic
 {
     public interface IAgeable
     {
-        long Age { get; set; }
-        long MaxAge { get; set; }
+        int Age { get; set; }
+        int MaxAge { get; set; }
     }
 }
