+++ first-world/Model/Animals.cs
@@ -1,3 +1,4 @@
+using System.Security.Cryptography;
 using System;
 using System.Numerics;
 using FirstWorld.Logic;
@@ -7,11 +8,11 @@ namespace FirstWorld.Model
     public class Animal : IObject, IAgeable, IBreedable
     {
         public Vector3 Position { set; get; }
-        public long Age { set; get; }
-        public long MaxAge { set; get; }
+        public int Age { set; get; }
+        public int MaxAge { set; get; }
         public int MaxBreedingNumber { set; get; }
-        public long BreedingAge { set; get; }
-        public long BreedingPeriod { set; get; }
+        public int BreedingAge { set; get; }
+        public int BreedingPeriod { set; get; }
         public float ChildSpreadingRadius { set; get; }
         public int MaxAgeVariation { set; get; }
         public int BreedingAgeVariation { set; get; }
@@ -21,9 +22,9 @@ namespace FirstWorld.Model
             Animal child = new Animal();
             child.Position = Position;
             child.Age = 0;
-            child.MaxAge = rand.Next((int)MaxAge - MaxAgeVariation,(int) MaxAge + MaxAgeVariation);
+            child.MaxAge = rand.Next(MaxAge - MaxAgeVariation, MaxAge + MaxAgeVariation);
             child.MaxBreedingNumber = MaxBreedingNumber;
-            child.BreedingAge = rand.Next((int)BreedingAge - BreedingAgeVariation, (int)BreedingAge + BreedingAgeVariation);
+            child.BreedingAge = rand.Next(BreedingAge - BreedingAgeVariation, BreedingAge + BreedingAgeVariation);
             child.BreedingPeriod = BreedingPeriod;
             child.ChildSpreadingRadius = ChildSpreadingRadius;
